clean up post layout
