+++ src/components/PostLayout/types.ts
@@ -1,4 +1,5 @@
 import { IGatsbyImageData } from 'gatsby-plugin-image'
+import React from 'react'
 
 export interface ITopic {
     name: string
@@ -14,6 +15,7 @@ export interface IContributor {
 }
 
 export interface IMenu {
+    icon?: React.ReactNode
     name: string
     url?: string
     children?: IMenu[]
@@ -58,12 +60,10 @@ export interface IProps {
         value: string
         depth: number
     }[]
-    children: React.ReactNode
     sidebar?: React.ReactNode
     contentWidth?: number | string
     questions?: React.ReactNode
     menu?: IMenu[]
-    article?: boolean
     title: string
     filePath?: string
     breadcrumb?: ICrumb[]
@@ -79,4 +79,8 @@ export interface IProps {
     }
     searchFilter?: string
     mobileMenu?: boolean
+    darkMode?: boolean
+    fullWidthContent?: boolean
+    setFullWidthContent: (fullWidth: boolean) => void
+    contentContainerClasses?: string
 }
