+++ src/templates/BlogPost.js
@@ -6,15 +6,13 @@ import { InlineCode } from 'components/InlineCode'
 import Layout from 'components/Layout'
 import Link from 'components/Link'
 import { H1, H2, H3, H4, H5, H6 } from 'components/MdxAnchorHeaders'
-import PostLayout, {
-    Contributor,
-    Contributors,
-    PageViews,
-    ShareLinks,
-    SidebarSection,
-    Text,
-    Topics,
-} from 'components/PostLayout'
+import PostLayout from 'components/PostLayout'
+import Text from 'components/PostLayout/Text'
+import Topics from 'components/PostLayout/Topics'
+import ShareLinks from 'components/PostLayout/ShareLinks'
+import SidebarSection from 'components/PostLayout/SidebarSection'
+import PageViews from 'components/PostLayout/PageViews'
+import Contributors, { Contributor } from 'components/PostLayout/Contributors'
 import { SEO } from 'components/seo'
 import { ZoomImage } from 'components/ZoomImage'
 import { graphql } from 'gatsby'
